pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "anglo-defects"
version = "23.7.211"
requires-python = ">=3.11"
dependencies = [
    'django==4.2.2',
    'python-dotenv==1.0.0',
    'whitenoise==6.5.0',
    'gunicorn==20.1.0',
    'dj-database-url==2.0.0',
    'django-lifecycle==1.0.0',
    'django-auditlog==2.3.0',
    'redis==4.5.5',
    'sentry_sdk==1.25.1',
    'django-import-export==3.2.0',
    'django-crispy-forms==2.0',
    'crispy-bootstrap5==0.7',
    'pillow==10.0.0',
]

[project.scripts]
"manage.py" = "manage.py:main"

[project.optional-dependencies]
dev = [
    "pip-tools",
    "black",
    "coverage",
]


[tool.setuptools]
packages = ['defects']
include-package-data = true

[tool.black]
target-version = ['py311']
line-length = 150

[tool.isort]
profile = "black"

[tool.coverage.run]
omit = [
    "*/migrations/*.py",
    "manage.py",
    "*/__init__.py",
]
